+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.7.0] - 2022-04-11
+- Added: support for [entity finder](https://github.com/heimrichhannot/contao-utils-bundle/blob/master/docs/commands/entity_finder.md)
+- Changed: raised heimrichhannot/contao-utils-bundle dependency to 2.213
+- Fixed: ConfigElementType exception if reader bundle is not installed
+- Fixed: removed travis config
+
 ## [2.6.0] - 2022-03-14
 - Changed: minimum php dependency is now php 7.4
 - Changed: updated coordinate SQL data types ([#20], [@qzminski])